replace int with float, delete absolute pathes

+++ src/render-components/Point.h
@@ -3,9 +3,9 @@
 #include <cstdint>
 
 struct Point {
-    uint32_t x;
-    uint32_t y;
+    float x;
+    float y;
 
     Point() : x(0), y(0) {}
-    Point(uint32_t x, uint32_t y) : x(x), y(y) {}
+    Point(float x, float y) : x(x), y(y) {}
 };
